Move TimeZones to the tools package
A time zone name is read wherever Legend touches one: in a date format string, on
a relational connection, in a plan parameter, in an H2 extension. Only the first
of those has anything to do with the Date core instance, and TimeZones sat under
coreinstance.primitive.date, beside the value that core instance holds. The class
is meant to be the one place any of those names is resolved, so that a name means
the same zone wherever it is read, and it belongs somewhere that says so.

tools.time is a new sub-package for time and date utilities that stand on their
own. The date classes already reach into tools for SafeAppendable and TextTools,
so the dependency runs the way it already ran.

Nothing has depended on the old location yet: TimeZones has not been released.
