+++ legend-pure-core/legend-pure-m4/src/test/java/org/finos/legend/pure/m4/tools/time/TestTimeZones.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.finos.legend.pure.m4.coreinstance.primitive.date;
+package org.finos.legend.pure.m4.tools.time;
 
 import org.junit.Assert;
 import org.junit.Test;
